requirements.txt
pytest>=3.0.7
Flask>=2.1.0
flask-cors>=3.0.10
google-cloud-pubsublite>=1.4.2